restructures helper functions
