+++ packages/db/functions/hacker.ts
@@ -1,19 +1,7 @@
-import { db, eq, sql } from ".";
-import { userCommonData } from "./schema";
-import { Hacker, User } from "./types";
+import { db, eq, sql } from "..";
+import { userCommonData } from "../schema";
+import { Hacker } from "../types";
 
-// USER
-
-const _getUser = db.query.userCommonData
-    .findFirst({
-        where: eq(userCommonData.clerkID, sql.placeholder('_clerkID')),
-    }).prepare("getUser");
-
-export function getUser(clerkID: string) : Promise<User | undefined> {
-    return _getUser.execute({_clerkID: clerkID});
-}
-
-// HACKER
 
 const _getHackerWithTeam = db.query.userCommonData
     .findFirst({
